requirements.txt
# Core Dependencies
langgraph==0.2.16
langchain==0.2.16
#langchain-openai==0.1.23
langchain-google-genai==2.1.12
langchain-core==0.2.38

# LLM and AI
#openai==1.45.0
google-genai==1.24.0

# HTTP and API
requests==2.32.3
httpx==0.27.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Environment Management
python-dotenv==1.0.1

# Logging and Monitoring
structlog==24.4.0

# Google Sheets Integration
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.144.0
gspread==6.1.2

# BuiltWith for tech stack data
builtwith==1.3.4

# Vector Store (for memory persistence)
chromadb==0.5.5
lancedb==0.13.0

# Utilities
pydantic==2.9.1
pydantic-settings==2.5.2
typing-extensions==4.12.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.8.0
flake8==7.1.1
mypy==1.11.2